Improved Security against malicious scripts
`expression` class in this code had a few security measures against malicious scripts.  
It had banned  `.`, `{` and `}`.  
Now there is an increase in security level, with the addition of `[` and `]` to this list for built-in security against JSFuck (JSFuck is an esoteric version of JavaScript which relies on `+`, `[`, `]`, `.`, `{` and `}`, and is used to pass malicious scripts, according to Wikipedia.)  
And `.` is removed from it, in order to be able to use decimal point literals.